Button that loads components
Pushing for working code that renders components into the main content window. I will add some comments in just a moment. As of right now the <App/> component will load a small blue button at the top left of the window that will change the splash screen to the trivia game and vice-versa, as proof of concept and reference.
I explect for the navbar to work, since it's a child of app, it will need to pass a "selection" prop to app for app to know what to render, which Donevan has some experience with.