Timer score implemented
The user's score for a given question is determined by how quickly they answer. As it is, there is no penalty for the first quarter of the timer duration, to give time to read the answers.
BUG, PROBABLY: I haven't tested with timers off but there may be some unset values causing math errors